Rename pkg constants to strings Implement Browser enum Rename ExceptionMessages to Messages Rename pkg runner to runners
