+++ src/test/java/jcucumberng/framework/enums/Browser.java
@@ -1,0 +1,5 @@
+package jcucumberng.framework.enums;
+
+public enum Browser {
+	CHROME32, CHROME32_NOHEAD, FF32, FF32_NOHEAD, FF64, FF64_NOHEAD, EDGE, IE32, IE64
+}
