+++ src/test/java/jcucumberng/framework/strings/Messages.java
@@ -1,6 +1,6 @@
-package jcucumberng.framework.constants;
+package jcucumberng.framework.strings;
 
-public abstract class ExceptionMessages {
+public abstract class Messages {
 	public static final String UNSUPPORTED_BROWSER = "See config.properties. Unsupported browser: ";
 	public static final String MISSING_ARGS = "No arguments found for arbitrary parameters. Length must not be equal to 0.";
 }
